requirements.txt
opencv_python==4.2.0.34
numpy==1.18.4
streamlit==0.62.1
Pillow==8.2.0
